Output a seed phrase as part of `verify`.
The verify command now takes the same --seed <bip39|mobile> option
as the create command. This will add a mnemonic phrase to the output
of `verify`. This phrase can be used to recreate the wallet.

It is important for the user to specify the correct seed type. The
same wallet will output different phrases when given different seed
types. The difference is the checksum bits at then end.

While it is possible to retrieve the phrase from a wallet  that was
originally imported from the mobile app, it is not possible to take
a cli-generated wallet (which has real 256-bits of entropy) and get
a 12-word phrase that the mobile wallet will accept.

TESTING:
I have not actually tested importing from mobile to CLI then back
to mobile.